Очередное изменение работы лоадера svg
